reorganize utility functions to helpers.py and place mappings in constants.py
