+++ weave/trace_server/opentelemetry/python_spans.py
@@ -34,17 +34,19 @@ from weave.trace_server import trace_server_interface as tsi
 from weave.trace_server.constants import MAX_DISPLAY_NAME_LENGTH, MAX_OP_NAME_LENGTH
 from weave.trace_server.opentelemetry.helpers import shorten_name
 
-from .attributes import (
+from weave.trace_server.opentelemetry.attributes import (
     SpanEvent,
     get_wandb_attributes,
     get_weave_attributes,
     get_weave_inputs,
     get_weave_outputs,
     get_weave_usage,
+)
+
+from weave.trace_server.opentelemetry.helpers import (
     to_json_serializable,
     unflatten_key_values,
 )
-
 
 class SpanKind(Enum):
     """Enum representing the span's kind."""
